XLog 프로파일 뷰의 상단에 SQL / API Call / OTHERS의 수행 시간 비율을 시각화 하여 표현 #144 - 피드백 반영하여 단독 컨텐츠로 변경 - 바의 색상 변경
