+++ src/components/Paper/XLog/Profiler/FrameProfile/FrameProfile.js
@@ -8,7 +8,7 @@ import numeral from "numeral";
 import moment from "moment";
 import {profileMetas} from "./meta"
 import {IdAbbr} from "../../../../../common/idAbbr";
-import ProfileRatioBar from './ProfileRatioBar';
+import TimeRatio from './TimeRatio';
 
 class FrameProfile extends Component {
 
@@ -453,13 +453,14 @@ class FrameProfile extends Component {
                 </div>
                 {
                     this.props.profile && 
-                    <ProfileRatioBar cpu={this.props.profile.cpu} 
+                    <TimeRatio cpu={this.props.profile.cpu} 
                         sql={this.props.profile.sqlTime} 
                         api={this.props.profile.apicallTime} 
                         elapsed={this.props.profile.elapsed}/>
                 }
                 
                 
+                <div className="sub-title">PROFILE STEP</div>
                 {/*<div className={"xlog-steps " + (this.props.wrap ? 'wrap' : '') + (this.props.narrow ? 'narrow' : '')}>*/}
                 <div className={"frame-graph " + (this.props.narrow ? 'narrow' : '')}>
                     <div ref="frameAxis" className="frame-axis"></div>
