+++ src/components/Paper/XLog/Profiler/FrameProfile/TimeRatio.css
@@ -1,5 +1,11 @@
 .ratio-bar {
   display: table;
+  margin: 10px 0 20px;
+  width: 100%;
+}
+
+.narrow .ratio-bar {
+  margin: 5px 0 20px;
 }
 
 .legend-row {
@@ -8,15 +14,6 @@
 
 .ratio-bar .ratio-legend {
   display: table-cell;
-}
-
-.ratio-bar .legend-header {
-  display: table-cell;
-}
-
-.ratio-bar .legend-header > div {
-  width: 78px;
-  height: 100%;
 }
 
 .ratio-bar .legend {
@@ -28,43 +25,34 @@
   content: " ";
   width: 10px;
   height: 10px;
-  border: 1px solid #333;
+  /* border: 1px solid #333; */
   margin-right: 3px;
 }
 
+.ratio-bar .legend .time {
+  font-size: 11px;
+}
+
 .ratio-bar .legend.cpu::before {
-  background-color: rgb(111, 169, 67);
+  background-color: #8b104e;
 }
 
 .ratio-bar .legend.sql::before {
-  background-color: rgb(113, 74, 209);
+  background-color: #104e8b;
 }
 
 .ratio-bar .legend.api::before {
-  background-color: rgb(87, 207, 243);
+  background-color: #4e8b10;
 }
 
 .ratio-bar .legend.other::before {
-  background-color: rgb(254, 250, 84);
+  background-color: #473C8B;
 }
 
 .ratio-row {
   display: table-row;
 }
 
-.ratio-bar .ratio-header {
-  display: table-cell;
-  padding: 10px;
-  border-radius: 3px;
-  background-color: #333;
-  color: white;
-  border: 1px solid #333;
-  font-size: 11px;
-}
-.ratio-bar .ratio-header > div {
-  width: 78px;
-  height: 100%;
-}
 .ratio-bar .ratio-body {
   display: table-cell;
   width: 100%;
@@ -73,17 +61,19 @@
   vertical-align: top;
 }
 
+.narrow .ratio-bar .ratio-body {
+  height: 20px;
+}
+
 .ratio-bar .ratio-body > div {
   display: table;
   width: 100%;
   height: 100%;
-  border: 1px solid #333;
 }
 
 .ratio-bar .ratio {
   display: table-cell;
   height: 100%;
-  border-right: 1px solid #333;
   padding: 10px 0;
   color: rgb(254,234, 51);
   overflow: hidden;
@@ -93,24 +83,27 @@
   max-width: 0;
 }
 
+.narrow .ratio-bar .ratio {
+  padding: 3px 0;
+}
+
 .ratio-bar .ratio:last-child {
   border-right: none;
 }
 
 .ratio-bar .ratio.cpu {
-  background-color: rgb(111, 169, 67);
+  background-color: #8b104e;
 }
 
 .ratio-bar .ratio.sql {
-  background-color: rgb(113, 74, 209);
+  background-color: #104e8b;
 }
 
 .ratio-bar .ratio.api {
-  background-color: rgb(87, 207, 243);
+  background-color: #4e8b10;
 }
 
 .ratio-bar .ratio.other {
-  background-color: rgb(254, 250, 84);
-  color: red;
+  background-color: #473C8B;
   width: 100%;
 }
